+++ CHANGELOG.md
@@ -13,6 +13,9 @@ You can find out backwards-compatibility policy [here](https://github.com/pyca/s
 <!-- changelog follows -->
 
 
+## [Unreleased](https://github.com/pyca/service-identity/compare/24.1.0...HEAD)
+
+
 ## [24.1.0](https://github.com/pyca/service-identity/compare/23.1.0...24.1.0) - 2024-01-14
 
 ### Changed